refactor/#26 애노테이션 추가, 패키지 명 변경
